[SPARK-56563][SQL][TESTS] Merge SparkPlanTest into QueryTest/QueryTestBase
### What changes were proposed in this pull request?

Followup to SPARK-55910 / SPARK-56560. Merge the helper methods from
`SparkPlanTest` (in `org.apache.spark.sql.execution`) into
`QueryTest`/`QueryTestBase` (in `org.apache.spark.sql`), and reduce
`SparkPlanTest` to a thin deprecated alias for backward compatibility.

Moved to `trait QueryTestBase` (non-anyfun instance helpers):
- `checkAnswer(DataFrame, SparkPlan => SparkPlan, Seq[Row], Boolean)`
- `checkAnswer2(DataFrame, DataFrame, (SparkPlan, SparkPlan) => SparkPlan, Seq[Row], Boolean)`
- `doCheckAnswer(Seq[DataFrame], Seq[SparkPlan] => SparkPlan, Seq[Row], Boolean)`
- `checkThatPlansAgree(DataFrame, SparkPlan => SparkPlan, SparkPlan => SparkPlan, Boolean)`

Moved to `object QueryTest` (static helpers):
- `checkAnswer(DataFrame, SparkPlan => SparkPlan, SparkPlan => SparkPlan, Boolean, SQLContext)`
- `checkAnswer(Seq[DataFrame], Seq[SparkPlan] => SparkPlan, Seq[Row], Boolean, SQLContext)`
- `executePlan(SparkPlan, SQLContext)`

`SparkPlanTest` becomes:
- `@deprecated private[sql] trait SparkPlanTest extends QueryTest`
- `@deprecated private[sql] object SparkPlanTest { /* thin delegates to QueryTest */ }`

Also removed the `= NUM_VALUES` default from the private
`ParquetVectorizedSuite.checkAnswer` to resolve a Scala
overload-with-default ambiguity against the new `QueryTestBase.checkAnswer`
(both call sites already pass `batchSize` explicitly, so behavior is
unchanged).

### Why are the changes needed?

Consolidates the test-helper surface into `QueryTest`/`QueryTestBase`,
matching the direction of SPARK-55910/SPARK-56560. New test suites can
extend `QueryTest` directly; existing `SparkPlanTest` subclasses and
callers of `SparkPlanTest.{checkAnswer, executePlan}` keep working via
the deprecated alias.

### Does this PR introduce _any_ user-facing change?

No, test-only.

### How was this patch tested?

Locally compiled `sql/Test/compile` and `hive/Test/compile`. Ran
`SortSuite` (255 tests), `ParquetVectorizedSuite` (25 tests), and
`SingleJoinSuite` (36 tests) — all pass.

### Was this patch authored or co-authored using generative AI tooling?

Generated-by: Claude Opus 4.7
