+++ sql/core/src/test/scala/org/apache/spark/sql/execution/SparkPlanTest.scala
@@ -17,237 +17,46 @@
 
 package org.apache.spark.sql.execution
 
-import scala.util.control.NonFatal
-
-import org.apache.spark.SparkFunSuite
-import org.apache.spark.sql.{DataFrame, QueryTest, Row, SparkSessionProvider, SQLContext}
-import org.apache.spark.sql.catalyst.analysis.UnresolvedAttribute
-import org.apache.spark.sql.catalyst.plans.logical.LocalRelation
-import org.apache.spark.sql.classic.ClassicConversions._
+import org.apache.spark.sql.{DataFrame, QueryTest, Row, SQLContext}
 
 /**
- * Base class for writing tests for individual physical operators. For an example of how this
- * class's test helper methods can be used, see [[SortSuite]].
+ * Kept as an empty alias of [[QueryTest]] for backward compatibility with existing subclasses.
+ * New test suites should extend [[QueryTest]] directly.
  */
-private[sql] abstract class SparkPlanTest extends SparkFunSuite with SparkSessionProvider {
+@deprecated("Use QueryTest directly instead.", "4.2.0")
+private[sql] trait SparkPlanTest extends QueryTest
+
+@deprecated("Use QueryTest directly instead.", "4.2.0")
+private[sql] object SparkPlanTest {
 
   /**
-   * Runs the plan and makes sure the answer matches the expected result.
-   * @param input the input data to be used.
-   * @param planFunction a function which accepts the input SparkPlan and uses it to instantiate
-   *                     the physical operator that's being tested.
-   * @param expectedAnswer the expected result in a [[Seq]] of [[Row]]s.
-   * @param sortAnswers if true, the answers will be sorted by their toString representations prior
-   *                    to being compared.
-   */
-  protected def checkAnswer(
-      input: DataFrame,
-      planFunction: SparkPlan => SparkPlan,
-      expectedAnswer: Seq[Row],
-      sortAnswers: Boolean = true): Unit = {
-    doCheckAnswer(
-      input :: Nil,
-      (plans: Seq[SparkPlan]) => planFunction(plans.head),
-      expectedAnswer,
-      sortAnswers)
-  }
-
-  /**
-   * Runs the plan and makes sure the answer matches the expected result.
-   * @param left the left input data to be used.
-   * @param right the right input data to be used.
-   * @param planFunction a function which accepts the input SparkPlan and uses it to instantiate
-   *                     the physical operator that's being tested.
-   * @param expectedAnswer the expected result in a [[Seq]] of [[Row]]s.
-   * @param sortAnswers if true, the answers will be sorted by their toString representations prior
-   *                    to being compared.
-   */
-  protected def checkAnswer2(
-      left: DataFrame,
-      right: DataFrame,
-      planFunction: (SparkPlan, SparkPlan) => SparkPlan,
-      expectedAnswer: Seq[Row],
-      sortAnswers: Boolean = true): Unit = {
-    doCheckAnswer(
-      left :: right :: Nil,
-      (plans: Seq[SparkPlan]) => planFunction(plans(0), plans(1)),
-      expectedAnswer,
-      sortAnswers)
-  }
-
-  /**
-   * Runs the plan and makes sure the answer matches the expected result.
-   * @param input the input data to be used.
-   * @param planFunction a function which accepts a sequence of input SparkPlans and uses them to
-   *                     instantiate the physical operator that's being tested.
-   * @param expectedAnswer the expected result in a [[Seq]] of [[Row]]s.
-   * @param sortAnswers if true, the answers will be sorted by their toString representations prior
-   *                    to being compared.
-   */
-  protected def doCheckAnswer(
-      input: Seq[DataFrame],
-      planFunction: Seq[SparkPlan] => SparkPlan,
-      expectedAnswer: Seq[Row],
-      sortAnswers: Boolean = true): Unit = {
-    SparkPlanTest
-      .checkAnswer(input, planFunction, expectedAnswer, sortAnswers, spark.sqlContext) match {
-        case Some(errorMessage) => fail(errorMessage)
-        case None =>
-    }
-  }
-
-  /**
-   * Runs the plan and makes sure the answer matches the result produced by a reference plan.
-   * @param input the input data to be used.
-   * @param planFunction a function which accepts the input SparkPlan and uses it to instantiate
-   *                     the physical operator that's being tested.
-   * @param expectedPlanFunction a function which accepts the input SparkPlan and uses it to
-   *                             instantiate a reference implementation of the physical operator
-   *                             that's being tested. The result of executing this plan will be
-   *                             treated as the source-of-truth for the test.
-   * @param sortAnswers if true, the answers will be sorted by their toString representations prior
-   *                    to being compared.
-   */
-  protected def checkThatPlansAgree(
-      input: DataFrame,
-      planFunction: SparkPlan => SparkPlan,
-      expectedPlanFunction: SparkPlan => SparkPlan,
-      sortAnswers: Boolean = true): Unit = {
-    SparkPlanTest.checkAnswer(
-        input, planFunction, expectedPlanFunction, sortAnswers, spark.sqlContext) match {
-      case Some(errorMessage) => fail(errorMessage)
-      case None =>
-    }
-  }
-}
-
-/**
- * Helper methods for writing tests of individual physical operators.
- */
-object SparkPlanTest {
-
-  /**
-   * Runs the plan and makes sure the answer matches the result produced by a reference plan.
-   * @param input the input data to be used.
-   * @param planFunction a function which accepts the input SparkPlan and uses it to instantiate
-   *                     the physical operator that's being tested.
-   * @param expectedPlanFunction a function which accepts the input SparkPlan and uses it to
-   *                             instantiate a reference implementation of the physical operator
-   *                             that's being tested. The result of executing this plan will be
-   *                             treated as the source-of-truth for the test.
+   * Kept as a thin alias of [[QueryTest.checkAnswer]] for backward compatibility.
+   * New callers should use [[QueryTest.checkAnswer]] directly.
    */
   def checkAnswer(
       input: DataFrame,
       planFunction: SparkPlan => SparkPlan,
       expectedPlanFunction: SparkPlan => SparkPlan,
       sortAnswers: Boolean,
-      spark: SQLContext): Option[String] = {
-
-    val outputPlan = planFunction(input.queryExecution.sparkPlan)
-    val expectedOutputPlan = expectedPlanFunction(input.queryExecution.sparkPlan)
-
-    val expectedAnswer: Seq[Row] = try {
-      executePlan(expectedOutputPlan, spark)
-    } catch {
-      case NonFatal(e) =>
-        val errorMessage =
-          s"""
-             | Exception thrown while executing Spark plan to calculate expected answer:
-             | $expectedOutputPlan
-             | == Exception ==
-             | $e
-             | ${org.apache.spark.sql.catalyst.util.stackTraceToString(e)}
-          """.stripMargin
-        return Some(errorMessage)
-    }
-
-    val actualAnswer: Seq[Row] = try {
-      executePlan(outputPlan, spark)
-    } catch {
-      case NonFatal(e) =>
-        val errorMessage =
-          s"""
-             | Exception thrown while executing Spark plan:
-             | $outputPlan
-             | == Exception ==
-             | $e
-             | ${org.apache.spark.sql.catalyst.util.stackTraceToString(e)}
-          """.stripMargin
-        return Some(errorMessage)
-    }
-
-    QueryTest.compareAnswers(actualAnswer, expectedAnswer, sortAnswers).map { errorMessage =>
-      s"""
-         | Results do not match.
-         | Actual result Spark plan:
-         | $outputPlan
-         | Expected result Spark plan:
-         | $expectedOutputPlan
-         | $errorMessage
-       """.stripMargin
-    }
-  }
+      spark: SQLContext): Option[String] =
+    QueryTest.checkAnswer(input, planFunction, expectedPlanFunction, sortAnswers, spark)
 
   /**
-   * Runs the plan and makes sure the answer matches the expected result.
-   * @param input the input data to be used.
-   * @param planFunction a function which accepts the input SparkPlan and uses it to instantiate
-   *                     the physical operator that's being tested.
-   * @param expectedAnswer the expected result in a [[Seq]] of [[Row]]s.
-   * @param sortAnswers if true, the answers will be sorted by their toString representations prior
-   *                    to being compared.
+   * Kept as a thin alias of [[QueryTest.checkAnswer]] for backward compatibility.
+   * New callers should use [[QueryTest.checkAnswer]] directly.
    */
   def checkAnswer(
       input: Seq[DataFrame],
       planFunction: Seq[SparkPlan] => SparkPlan,
       expectedAnswer: Seq[Row],
       sortAnswers: Boolean,
-      spark: SQLContext): Option[String] = {
-
-    val outputPlan = planFunction(input.map(_.queryExecution.sparkPlan))
-
-    val sparkAnswer: Seq[Row] = try {
-      executePlan(outputPlan, spark)
-    } catch {
-      case NonFatal(e) =>
-        val errorMessage =
-          s"""
-             | Exception thrown while executing Spark plan:
-             | $outputPlan
-             | == Exception ==
-             | $e
-             | ${org.apache.spark.sql.catalyst.util.stackTraceToString(e)}
-          """.stripMargin
-        return Some(errorMessage)
-    }
-
-    QueryTest.compareAnswers(sparkAnswer, expectedAnswer, sortAnswers).map { errorMessage =>
-      s"""
-         | Results do not match for Spark plan:
-         | $outputPlan
-         | $errorMessage
-       """.stripMargin
-    }
-  }
+      spark: SQLContext): Option[String] =
+    QueryTest.checkAnswer(input, planFunction, expectedAnswer, sortAnswers, spark)
 
   /**
-   * Runs the plan
-   * @param outputPlan SparkPlan to be executed
-   * @param spark SqlContext used for execution of the plan
+   * Kept as a thin alias of [[QueryTest.executePlan]] for backward compatibility.
+   * New callers should use [[QueryTest.executePlan]] directly.
    */
-  def executePlan(outputPlan: SparkPlan, spark: SQLContext): Seq[Row] = {
-    val execution = new QueryExecution(spark.sparkSession, LocalRelation(Nil)) {
-      override lazy val sparkPlan: SparkPlan = outputPlan transform {
-        case plan: SparkPlan =>
-          val inputMap = plan.children.flatMap(_.output).map(a => (a.name, a)).toMap
-          plan transformExpressions {
-            case UnresolvedAttribute(Seq(u)) =>
-              inputMap.getOrElse(u,
-                sys.error(s"Invalid Test: Cannot resolve $u given input $inputMap"))
-          }
-      }
-    }
-    execution.executedPlan.executeCollectPublic().toSeq
-  }
+  def executePlan(outputPlan: SparkPlan, spark: SQLContext): Seq[Row] =
+    QueryTest.executePlan(outputPlan, spark)
 }
-
